python-service/requirements.txt
# Flask核心框架
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0

# 数据库
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# 配置和环境管理
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Nacos服务发现
nacos-sdk-python==0.1.12
requests==2.31.0

# API文档
flask-swagger-ui==4.11.1
flasgger==0.9.7.1

# 日志和监控
structlog==23.2.0
prometheus-client==0.19.0

# 工具库
celery==5.3.4
gunicorn==21.2.0

# 开发和测试
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1 